+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.1] - 2025-06-09
+
+### Changed
+
+- Instead of recursively applying patches, now it will apply them one by one
+  in chain to reduce memory usage [AdguardBrowserExtension#3230].
+
+[1.1.1]: https://github.com/AdguardTeam/DiffBuilder/compare/v1.1.0...v1.1.1
+[AdguardBrowserExtension#3230]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3230
+
 ## [1.1.0] - 2025-03-13
 
 ### Changed